% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clientsideHtmlPassword.R
\name{protect}
\alias{protect}
\title{Wrapper for Clientside HTML Password Protect by Max Laumeister and Zoltan Galli.}
\usage{
protect(file, passphrase, template = NULL)
}
\arguments{
\item{file}{Path to file to password protected}

\item{passphrase}{Single character string. Passphrase for open the file}

\item{template}{Template HTML file}
}
\value{
Character string of protected file.
}
\description{
Wrapper for Clientside HTML Password Protect by Max Laumeister and Zoltan Galli.
}
\examples{
\dontrun{
file <- system.file("sample", "hello.html", package = "clientsideHtmlProtect")
protected <- protect(file, "abc")
writeLines(protected, "protect.html")
file.remove("protect.html")
}
}
